Separated CC and non-CC dashboards
